+++ test/unit/import.helper.test.ts
@@ -1,11 +1,10 @@
 import { assert, expect } from "chai";
 import path from "path";
 import proxyquire from "proxyquire";
-import { SinonSpy } from "sinon";
 import sinon from "sinon";
-import { ObjectReadableMock, ObjectWritableMock } from "stream-mock";
-import { FileHelper, GitHelper, ImportHelper, ProjectHelper } from "../../helper";
-import { ICsvRow, IRecord } from "../../interfaces";
+import { ObjectReadableMock } from "stream-mock";
+import { FileHelper, GitHelper, ImportHelper } from "../../helper";
+import { IRecord } from "../../interfaces";
 
 const sandboxDir: string = "./sandbox";
 const configDir: string = path.join(sandboxDir, ".git-time-tracker");
@@ -40,7 +39,6 @@ const expectedEmptyArray: IRecord[] = [];
 
 describe("ImportHelper", () => {
   let mockedFileHelper: FileHelper;
-  let mockedProjectHelper: ProjectHelper;
   let mockedGitHelper: GitHelper;
 
   before(() => {
@@ -57,11 +55,10 @@ describe("ImportHelper", () => {
 
     mockedFileHelper = new fileProxy.FileHelper(configDir, configFileName, timerFileName, projectsDir);
     mockedGitHelper = new gitProxy.GitHelper(configDir, mockedFileHelper);
-    mockedProjectHelper = new projectProxy.ProjectHelper(mockedGitHelper, mockedFileHelper);
   });
 
   it("should create instance", async () => {
-    const importHelper: ImportHelper = new ImportHelper(mockedProjectHelper);
+    const importHelper: ImportHelper = new ImportHelper();
     expect(importHelper).to.be.instanceOf(ImportHelper);
   });
 
@@ -72,7 +69,7 @@ describe("ImportHelper", () => {
       },
     });
 
-    const instance: ImportHelper = new fileProxy.ImportHelper(mockedProjectHelper);
+    const instance: ImportHelper = new fileProxy.ImportHelper();
     const result: IRecord[] = await instance.importCsv(csvFilePath);
 
     assert.isArray(result);
@@ -86,7 +83,7 @@ describe("ImportHelper", () => {
       },
     });
 
-    const instance: ImportHelper = new fileProxy.ImportHelper(mockedProjectHelper);
+    const instance: ImportHelper = new fileProxy.ImportHelper();
     const result: IRecord[] = await instance.importCsv(csvFilePath);
 
     assert.isArray(result);
@@ -100,7 +97,7 @@ describe("ImportHelper", () => {
       },
     });
 
-    const instance: ImportHelper = new fileProxy.ImportHelper(mockedProjectHelper);
+    const instance: ImportHelper = new fileProxy.ImportHelper();
     const result: IRecord[] = await instance.importCsv(csvFilePath);
 
     assert.isArray(result);
@@ -114,7 +111,7 @@ describe("ImportHelper", () => {
       },
     });
 
-    const instance: ImportHelper = new fileProxy.ImportHelper(mockedProjectHelper);
+    const instance: ImportHelper = new fileProxy.ImportHelper();
     const result: IRecord[] = await instance.importCsv(csvFilePath);
 
     assert.isArray(result);
@@ -128,7 +125,7 @@ describe("ImportHelper", () => {
       },
     });
 
-    const instance: ImportHelper = new fileProxy.ImportHelper(mockedProjectHelper);
+    const instance: ImportHelper = new fileProxy.ImportHelper();
     const result: IRecord[] = await instance.importCsv(csvFilePath);
 
     assert.isArray(result);
@@ -142,7 +139,7 @@ describe("ImportHelper", () => {
       },
     });
 
-    const instance: ImportHelper = new fileProxy.ImportHelper(mockedProjectHelper);
+    const instance: ImportHelper = new fileProxy.ImportHelper();
     const result: IRecord[] = await instance.importCsv(csvFilePath);
 
     assert.isArray(result);
@@ -156,7 +153,7 @@ describe("ImportHelper", () => {
       },
     });
 
-    const instance: ImportHelper = new fileProxy.ImportHelper(mockedProjectHelper);
+    const instance: ImportHelper = new fileProxy.ImportHelper();
     const result: IRecord[] = await instance.importCsv(csvFilePath);
 
     assert.isArray(result);
@@ -170,7 +167,7 @@ describe("ImportHelper", () => {
       },
     });
 
-    const instance: ImportHelper = new fileProxy.ImportHelper(mockedProjectHelper);
+    const instance: ImportHelper = new fileProxy.ImportHelper();
     const result: IRecord[] = await instance.importCsv(csvFilePath);
 
     assert.isArray(result);
