+++ test/unit/question.helper.test.ts
@@ -1,4 +1,4 @@
-import { assert, expect } from "chai";
+import { expect } from "chai";
 import proxyquire from "proxyquire";
 import sinon from "sinon";
 import { QuestionHelper } from "../../helper";
